+++ packages/terra-menu/tests/wdio/menu-item-spec.js
@@ -31,6 +31,36 @@ Terra.describeViewports('Menu Item', ['medium'], () => {
     });
   });
 
+  describe('Menu navigation to nested sub menu when clicked', () => {
+    it('retains selected item on nested menus while navigating back and forth between them', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/menu/menu/nested-sub-menu-selectable');
+      $('#default-button').click();
+      $('[class*=subMenu1]').click();
+      expect($('[class*=testNestedItem1]').isFocused()).toBeTruthy();
+      $('[class*=testNestedItem1]').click();
+      $('[role="button"][aria-label="Back"]').click();
+      $('[class*=subMenu1]').click();
+      expect($('li:first-child[aria-checked="true"][role="menuitemcheckbox"]').isFocused()).toBeTruthy();
+      Terra.validates.element('maintained selection after navigate to sub menu', { selector: '#root' });
+    });
+  });
+
+  describe('Menu-Selectable with Varying Items', () => {
+    it('displays a selectable menu with varying items', () => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/menu/menu/selectable-and-unselectable-items-menu');
+      $('#default-button').click();
+      Terra.validates.element('with varying items', { selector: '#root' });
+    });
+
+    it('selects an item and maintains selection after menu has been reopened', () => {
+      $('.TestSelectableItem').click();
+      expect($('#default-button').isFocused()).toBeTruthy();
+      $('#default-button').click();
+      expect($('li:first-child[role="menuitem"]').isFocused()).toBeTruthy();
+      Terra.validates.element('maintained selection after reopen', { selector: '#root' });
+    });
+  });
+
   describe('Menu Item-Icon', () => {
     it('displays Menu Items with icon', () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/menu/menu-item/menu-items-with-icons');