+++ packages/terra-form-validation/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.80.0 - (July 4, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.79.0 - (June 12, 2023)
 
 * Changed