+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,16 +1,30 @@
 # Changelog
 
 ## Unreleased
+
 * Changed
+  * Updated example for `terra-menu`.
+
+* Added
+  * Added alpha release disclaimer for `terra-worklist-data-grid`.
+
+## 1.24.0 - (July 4, 2023)
+
+* Updated
+  * Updated examples and tests for `terra-menu`.
+
+* Changed
+  * Updated example for `terra-popup`.
   * Updated examples and tests for `terra-worklist-data-grid` to remove row height
   * Updated examples and tests for `terra-worklist-data-grid` for blank cell accessibility
   * Updated examples and tests for `terra-worklist-data-grid` to show masked cells and hover states
   * Updated tests for `terra-worklist-data-grid` to account for row height and column width being honored
+  * Updated examples and tests for `terra-worklist-data-grid` to show column sort and error indicators
 
 ## 1.23.0 - (June 12, 2023)
 
 * Changed
-  * Minor dependency version bump
+  * Minor dependency version bump.
 
 ## 1.22.0 - (June 5, 2023)
 
@@ -26,6 +40,7 @@
 * Added
   * Added documentation and tests for `terra-worklist-data-grid`.
   * Added example and test for masked cell feature for `terra-worklist-data-grid`.
+
 
 ## 1.21.0 - (May 11, 2023)