+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -2,13 +2,18 @@
 
 ## Unreleased
 * Changed
-  * Removed ability to set variable row heights
+  * Updated data grid to honor the column and row widths specified by consumers.
+  
+## 0.2.0 - (July 4, 2023)
+
+* Changed
+  * Removed ability to set variable row heights.
   * Updated styling on cell focus.
-  * Updates data grid to honor the column and row widths specified by consumers
 * Added
   * Added masked cell support.
   * Added hover state for selectable cells.
   * Added blank cell support.
+  * Added ability to allow consumers to specify the sort indicator and error state for columns.
 
 ## 0.1
   * Initial stable release.